Move Errno into telemetry
Following on #50, this change moves the `Errno` types into `telemetry`
so that metrics implementation can be moved out of the `metrics` module.
