+++ src/telemetry/metrics/labels/mod.rs
@@ -9,11 +9,8 @@ use http;
 
 use ctx;
 use conditional::Conditional;
-use telemetry::event;
+use telemetry::{event, Errno};
 use transport::tls;
-
-mod errno;
-pub use self::errno::Errno;
 
 #[derive(Clone, Debug, Eq, PartialEq, Hash)]
 pub struct RequestLabels {
@@ -434,24 +431,5 @@ impl fmt::Display for tls::ReasonForNoIdentity {
             tls::ReasonForNoIdentity::NotImplementedForMetrics =>
                 f.pad("not_implemented_for_metrics"),
         }
-    }
-}
-
-
-#[cfg(target_os="windows")]
-pub struct Errno(i32);
-
-#[cfg(target_os="windows")]
-impl From<i32> for Errno {
-    fn from(code: i32) -> Self {
-        Errno(code)
-    }
-}
-
-#[cfg(target_os="windows")]
-#[derive(Copy, Clone, Debug, Eq, PartialEq, Hash)]
-impl fmt::Display for Errno {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        fmt::Display.fmt(self.0, f)
     }
 }
